+++ node/src/types/kernel/api/filetree/createDocWithMd/index.d.ts
@@ -17,9 +17,7 @@
 
 
 /* directories */
-export * as createDocWithMd from "./createDocWithMd";
 
 /* files */
-export * as listDocsByPath from "./listDocsByPath";
-export * as renameDoc from "./renameDoc";
-export * as searchDocs from "./searchDocs";
+export * from "./payload";
+export * from "./response";